requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
coverage==5.3.1
discord.py==1.6.0
idna==2.10
importlib-metadata==3.4.0
iniconfig==1.1.1
liquipediapy==2.1.0
lxml==4.6.2
multidict==5.1.0
packaging==20.8
pluggy==0.13.1
pprintpp==0.4.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.11.1
pytest-mock==3.5.1
python-dateutil==2.8.1
pytz==2020.5
pytz-convert==0.2.9
requests==2.25.1
six==1.15.0
soupsieve==2.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
zipp==3.4.0
